+++ comb_spec_searcher/comb_spec_searcher.py
@@ -95,7 +95,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             - `debug`: if True every rule found will be sanity checked and logged
               to logging.DEBUG
             - `function_kwargs` are passed to the call method of strategies
-            - `logger_kwargs` are passed to the logger when logging
         """
         self.strategy_pack = strategy_pack
         self.debug = kwargs.get("debug", False)
@@ -103,12 +102,10 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         if self.debug:
             logzero.loglevel(logging.DEBUG, True)
         self.kwargs = kwargs.get("function_kwargs", dict())
-        self.logger_kwargs = kwargs.get("logger_kwargs", {"processname": "runner"})
 
         self.func_times: Dict[str, float] = defaultdict(float)
         self.func_calls: Dict[str, int] = defaultdict(int)
 
-        self.kwargs["logger"] = self.logger_kwargs
         self.kwargs["symmetry"] = bool(strategy_pack.symmetries)
 
         self.classdb = ClassDB[CombinatorialClassType](type(start_class))
@@ -235,7 +232,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             "Expanding label %s with %s",
             label,
             strategy_generator,
-            extra=self.logger_kwargs,
         )
         if label is None:
             label = self.classdb.get_label(comb_class)
@@ -251,7 +247,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                     "combinatorial class when applied to %r",
                     str(rule).split(" ")[1],
                     comb_class,
-                    extra=self.logger_kwargs,
                 )
                 continue
             end_labels = [self.classdb.get_label(child) for child in children]
@@ -268,7 +263,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                     start_label,
                     tuple(end_labels),
                     rule,
-                    extra=self.logger_kwargs,
                 )
                 try:
                     n = 4
@@ -318,9 +312,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             # Only applying is_empty check to comb classes that are
             # possibly empty.
             if rule.possibly_empty and self.is_empty(comb_class, child_label):
-                logger.debug(
-                    "Label %s is empty.", child_label, extra=self.logger_kwargs
-                )
+                logger.debug("Label %s is empty.", child_label)
                 continue
             if rule.workable:
                 self._add_to_queue(child_label)
@@ -546,7 +538,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         found_string += (
             f"Specification found has {specification.number_of_rules()} rules"
         )
-        logger.info(found_string, extra=self.logger_kwargs)
+        logger.info(found_string)
 
     def _log_status(self, start_time: float, status_update: int) -> None:
         time_taken = time.time() - start_time
@@ -565,7 +557,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 " -- next elaborate status update in "
                 f"{timedelta(seconds=next_elaborate)} --\n"
             )
-        logger.info(status, extra=self.logger_kwargs)
+        logger.info(status)
 
     def auto_search(self, **kwargs) -> CombinatorialSpecification:
         """
@@ -598,7 +590,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                     "Percentage not between 0 and 100, so assuming 1%"
                     " search percentage."
                 ),
-                extra=self.logger_kwargs,
             )
             perc = 1
         status_update = kwargs.get("status_update", None)
@@ -608,7 +599,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             time.strftime("%a, %d %b %Y %H:%M:%S", time.gmtime())
         )
         start_string += self.run_information()
-        logger.info(start_string, extra=self.logger_kwargs)
+        logger.info(start_string)
 
         max_expansion_time = 0
         expanding = True
@@ -619,7 +610,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             )
 
             spec_search_start = time.time()
-            logger.debug("Searching for specification.", extra=self.logger_kwargs)
+            logger.debug("Searching for specification.")
             specification = self.get_specification(
                 smallest=kwargs.get("smallest", False),
                 minimization_time_limit=0.01 * (time.time() - auto_search_start),
@@ -627,7 +618,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             if specification is not None:
                 self._log_spec_found(specification, auto_search_start)
                 return specification
-            logger.debug("No specification found.", extra=self.logger_kwargs)
+            logger.debug("No specification found.")
             if max_time is not None:
                 if time.time() - auto_search_start > max_time:
                     raise ExceededMaxtimeError(
@@ -641,7 +632,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             logger.debug(
                 "Will expand for %s seconds.",
                 round(max_expansion_time, 2),
-                extra=self.logger_kwargs,
             )
 
     def _auto_search_rules(self) -> Specification:
@@ -702,7 +692,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 status_start = time.time()
         else:
             expanding = False
-            logger.info("No more classes to expand.", extra=self.logger_kwargs)
+            logger.info("No more classes to expand.")
         return expanding, status_start
 
     @cssmethodtimer("get specification")
@@ -722,7 +712,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             return None
         start_class = self.classdb.get_class(self.start_label)
         strategies, comb_class_eqv_paths = spec
-        logger.info("Creating a specification.", extra=self.logger_kwargs)
+        logger.info("Creating a specification.")
         return CombinatorialSpecification(start_class, strategies, comb_class_eqv_paths)
 
     @cssmethodtimer("get specification")